+++ CHANGELOG.md
@@ -6,4 +6,6 @@
 
 ### Changes
 
+* Replaced method `find_named` with `find_by_name`.
+
 ### Bugs fixed